Add support for alternative operator representations
This adds support for the alternative operator representations in unary,
binary, and assignment expressions, as well as when using in explicit
operator calls and overload declarations/definitions.

This does _not_ support full token substitution the preprocessor
performs (these alternate tokens can be substituted _anywhere_ the
tokens they represent appear, but in this grammar they are only
supported in the above expressions).

This also does _not_ add support for alternate token representations of
the brace tokens ('{', '}', '[', ']') or preprocessor directives
('#', '##')